BM328-Tasarim-Calismasi-II/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
decorator==5.1.1
Django==4.2
django-autoslug==1.9.9
django-crispy-forms==2.0
django-extensions==3.2.1
django-tinymce==3.6.1
executing==1.2.0
idna==3.4
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-slugify==8.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
stripe==5.4.0
text-unidecode==1.3
traitlets==5.9.0
urllib3==2.0.2
wcwidth==0.2.6
